Move lower level Affine logic into odc.geo.math
This is to decouple _overlap from the rest, so that
we can enable types without import loops.
