+++ odc/geo/_geobox.py
@@ -18,10 +18,8 @@ from ._geom import (
     bbox_union,
     polygon_from_transform,
 )
-from ._overlap import is_affine_st
-from .math import is_almost_int
-
 from ._roi import roi_normalise, roi_shape
+from .math import is_affine_st, is_almost_int
 
 Coordinate = namedtuple("Coordinate", ("values", "units", "resolution"))
