+++ odc/geo/_overlap.py
@@ -11,55 +11,7 @@ from affine import Affine
 from numpy import linalg
 
 from ._roi import gbox_boundary, roi_boundary, roi_center, roi_from_points, roi_is_empty
-from .math import maybe_int, snap_scale, split_float
-
-# This is numeric code, short names make sense in this context, so disabling
-# "invalid name" checks for the whole file
-# pylint: disable=invalid-name
-
-
-def apply_affine(
-    A: Affine, x: np.ndarray, y: np.ndarray
-) -> Tuple[np.ndarray, np.ndarray]:
-    """
-    broadcast A*(x_i, y_i) across all elements of x/y arrays in any shape (usually 2d image)
-    """
-
-    shape = x.shape
-
-    A = np.asarray(A).reshape(3, 3)
-    t = A[:2, -1].reshape((2, 1))
-    A = A[:2, :2]
-
-    x, y = A @ np.vstack([x.ravel(), y.ravel()]) + t
-    x, y = (a.reshape(shape) for a in (x, y))
-    return (x, y)
-
-
-def split_translation(t):
-    """
-    Split translation into pixel aligned and sub-pixel components.
-
-    Subpixel translation is guaranteed to be in [-0.5, +0.5] range.
-
-    >  x + t = x + t_whole + t_subpix
-
-    :param t: (float, float)
-
-    :returns: (t_whole: (float, float), t_subpix: (float, float))
-    """
-
-    _tt = [split_float(x) for x in t]
-    return tuple(t[0] for t in _tt), tuple(t[1] for t in _tt)
-
-
-def is_affine_st(A, tol=1e-10):
-    """
-    True if Affine transform has scale and translation components only.
-    """
-    (_, wx, _, wy, _, _, *_) = A
-
-    return abs(wx) < tol and abs(wy) < tol
+from .math import is_affine_st, maybe_int, snap_scale
 
 
 def decompose_rws(A):
