+++ odc/geo/testutils/geom.py
@@ -7,7 +7,8 @@ from typing import Callable, Tuple, Union
 import numpy as np
 from affine import Affine
 
-from odc.geo import CRS, GeoBox, GridSpec, apply_affine
+from odc.geo import CRS, GeoBox, GridSpec
+from odc.geo.math import apply_affine
 
 # pylint: disable=invalid-name
